requirements.txt
numpy
scipy
ase
dscribe
mace-torch
cuequivariance-torch
matplotlib
pymatgen
pyscf
scikit-learn
spglib
seekpath
phonopy


